dynamically load new switch config
