+++ pkg/commands/commands.go
@@ -6,9 +6,10 @@ import (
 
 	"github.com/dhowlett99/dmxlights/pkg/common"
 	"github.com/dhowlett99/dmxlights/pkg/config"
+	"github.com/dhowlett99/dmxlights/pkg/fixture"
 )
 
-const debug = false
+const debug = true
 const beatDebug = false
 
 // listenCommandChannelAndWait listens on channel for instructions or timeout and go to next step of sequence.
@@ -377,10 +378,22 @@ func ListenCommandChannelAndWait(mySequenceNumber int, currentSpeed time.Duratio
 
 	// Clear switch positions for this sequence
 	case common.ClearAllSwitchPositions:
+		if debug {
+			fmt.Printf("%d: Command ClearAllSwitchPositions n", mySequenceNumber)
+		}
 		// Loop through all the switchies.
 		for X := 0; X < len(sequence.Switches); X++ {
 			sequence.Switches[X].CurrentState = 0
 		}
+		sequence.PlaySwitchOnce = true
+		return sequence
+
+	case common.ResetAllSwitchPositions:
+		const FIXTURES_CONFIG = 0
+		if debug {
+			fmt.Printf("%d: Command ResetAllSwitchPositions n", mySequenceNumber)
+		}
+		sequence.Switches = LoadSwitchConfiguration(mySequenceNumber, command.Args[FIXTURES_CONFIG].Value.(*fixture.Fixtures))
 		sequence.PlaySwitchOnce = true
 		return sequence
 
@@ -566,4 +579,65 @@ func SetSpeed(commandSpeed int) (Speed time.Duration) {
 		Speed = 75
 	}
 	return Speed * time.Millisecond
+}
+
+func LoadSwitchConfiguration(mySequenceNumber int, fixturesConfig *fixture.Fixtures) []common.Switch {
+
+	if debug {
+		fmt.Printf("Load switch data\n")
+	}
+
+	// A new group of switches.
+	newSwitchList := []common.Switch{}
+	for _, fixture := range fixturesConfig.Fixtures {
+		if fixture.Group == mySequenceNumber+1 {
+			// find switch data.
+			newSwitch := common.Switch{}
+			newSwitch.Name = fixture.Name
+			newSwitch.Label = fixture.Label
+			newSwitch.Number = fixture.Number
+			newSwitch.Description = fixture.Description
+			newSwitch.UseFixture = fixture.UseFixture
+
+			newSwitch.States = []common.State{}
+			for _, state := range fixture.States {
+				newState := common.State{}
+				newState.Name = state.Name
+				newState.Number = state.Number
+				newState.Label = state.Label
+				newState.ButtonColor = state.ButtonColor
+				newState.Flash = state.Flash
+
+				// Copy values.
+				newState.Values = []common.Value{}
+				for _, value := range state.Values {
+					newValue := common.Value{}
+					newValue.Channel = value.Channel
+					newValue.Setting = value.Setting
+					newState.Values = append(newState.Values, newValue)
+				}
+
+				// Copy actions.
+				newState.Actions = []common.Action{}
+				for _, action := range state.Actions {
+					newAction := common.Action{}
+					newAction.Name = action.Name
+					newAction.Colors = action.Colors
+					newAction.Mode = action.Mode
+					newAction.Fade = action.Fade
+					newAction.Speed = action.Speed
+					newAction.Rotate = action.Rotate
+					newAction.Music = action.Music
+					newAction.Program = action.Program
+					newState.Actions = append(newState.Actions, newAction)
+				}
+
+				newSwitch.States = append(newSwitch.States, newState)
+			}
+			// Add new switch to the list.
+			newSwitchList = append(newSwitchList, newSwitch)
+		}
+	}
+
+	return newSwitchList
 }
